python_toolbox/requirements.txt
# Python工具箱依赖列表
# 基础依赖
pillow>=8.0.0  # 用于图像处理功能
requests>=2.25.0  # 用于HTTP请求功能
PyQt5>=5.15.0  # 用于GUI界面功能